ideal

+++ IDEAL/IDEAL-PARENT/IDEAL-COMMON/src/main/java/cn/ideal/common/mapper/AccountUserMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface AccountUserMapper {
-    int countByExample(AccountUserExample example);
+    long countByExample(AccountUserExample example);
 
     int deleteByExample(AccountUserExample example);
